Move ValueWithRecordId to sdk.values, annotated
